Feat: Peek 상세 조회 시 거리 반환 추가

+++ backend/src/main/java/com/vvs/peekpick/peek/dto/ResponsePeekListDto.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class ResponsePeekListDto {
     private Long peekId;
-    private Double distance;
+    private int distance;
     private boolean special;
     private boolean viewed;
 }
